refactor: use common plus/minus/intersect instead of generics
