+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/MultiSet.kt
@@ -1,5 +1,10 @@
 package xyz.lbres.kotlinutils.set.multiset
 
+import xyz.lbres.kotlinutils.set.multiset.utils.CountsMap
+import xyz.lbres.kotlinutils.set.multiset.utils.performIntersect
+import xyz.lbres.kotlinutils.set.multiset.utils.performMinus
+import xyz.lbres.kotlinutils.set.multiset.utils.performPlus
+
 /**
  * Interface for set that allows multiple occurrences of a value.
  * The interface supports only read access to the values.
@@ -55,7 +60,9 @@ interface MultiSet<E> : Collection<E> {
          * @param multiSet2 [MultiSet]<E>: second MultiSet in addition
          * @return [MultiSet]<E>: MultiSet containing all values from both sets
          */
-        fun <E> genericPlus(multiSet1: MultiSet<E>, multiSet2: MultiSet<E>): MultiSet<E> = genericMultiSetPlus(multiSet1, multiSet2)
+        fun <E> genericPlus(multiSet1: MultiSet<E>, multiSet2: MultiSet<E>): MultiSet<E> {
+            return performPlus(CountsMap.from(multiSet1), multiSet2)
+        }
 
         /**
          * [minus] implementation that can be used with any [MultiSet] implementations.
@@ -65,7 +72,9 @@ interface MultiSet<E> : Collection<E> {
          * @param multiSet2 [MultiSet]<E>: second MultiSet in subtraction
          * @return [MultiSet]<E>: MultiSet containing the items in the first MultiSet but not the second
          */
-        fun <E> genericMinus(multiSet1: MultiSet<E>, multiSet2: MultiSet<E>): MultiSet<E> = genericMultiSetMinus(multiSet1, multiSet2)
+        fun <E> genericMinus(multiSet1: MultiSet<E>, multiSet2: MultiSet<E>): MultiSet<E> {
+            return performMinus(CountsMap.from(multiSet1), multiSet2)
+        }
 
         /**
          * [minus] implementation that can be used with any [MultiSet] implementations.
@@ -75,6 +84,8 @@ interface MultiSet<E> : Collection<E> {
          * @param multiSet2 [MultiSet]<E>: second MultiSet in intersect
          * @return [MultiSet]<E>: MultiSet containing only values that are in both sets
          */
-        fun <E> genericIntersect(multiSet1: MultiSet<E>, multiSet2: MultiSet<E>): MultiSet<E> = genericMultiSetIntersect(multiSet1, multiSet2)
+        fun <E> genericIntersect(multiSet1: MultiSet<E>, multiSet2: MultiSet<E>): MultiSet<E> {
+            return performIntersect(CountsMap.from(multiSet1), multiSet2)
+        }
     }
 }
